Remove redundancy in NoteCommitConfig/Chip
