+++ src/circuit/circuit_vanilla.rs
@@ -42,11 +42,10 @@ use super::{
         add_chip::{self, AddChip, AddConfig},
         assign_free_advice, AddInstruction,
     },
+    note_commit::{NoteCommitChip, NoteCommitConfig},
     Circuit, OrchardCircuit, ANCHOR, CMX, CV_NET_X, CV_NET_Y, ENABLE_OUTPUT, ENABLE_SPEND, NF_OLD,
     RK_X, RK_Y,
 };
-
-use self::note_commit::{NoteCommitChip, NoteCommitConfig};
 
 mod gadget;
 mod note_commit;
@@ -240,12 +239,12 @@ impl OrchardCircuit for OrchardVanilla {
         // Configuration to handle decomposition and canonicity checking
         // for NoteCommit_old.
         let old_note_commit_config =
-            NoteCommitChip::configure(meta, advices, sinsemilla_config_1.clone());
+            NoteCommitChip::configure(meta, advices, sinsemilla_config_1.clone(), false);
 
         // Configuration to handle decomposition and canonicity checking
         // for NoteCommit_new.
         let new_note_commit_config =
-            NoteCommitChip::configure(meta, advices, sinsemilla_config_2.clone());
+            NoteCommitChip::configure(meta, advices, sinsemilla_config_2.clone(), false);
 
         Config {
             primary,
