+++ src/circuit/circuit_vanilla/gadget.rs
@@ -2,8 +2,11 @@
 
 use pasta_curves::pallas;
 
-use super::{add_chip, commit_ivk::CommitIvkChip, note_commit::NoteCommitChip, AddInstruction};
-use crate::constants::{OrchardCommitDomains, OrchardFixedBases, OrchardHashDomains};
+use super::{add_chip, commit_ivk::CommitIvkChip, AddInstruction};
+use crate::{
+    circuit::note_commit::NoteCommitChip,
+    constants::{OrchardCommitDomains, OrchardFixedBases, OrchardHashDomains},
+};
 use halo2_gadgets::{
     ecc::{
         chip::{EccChip, EccPoint},
