+++ src/circuit/circuit_vanilla/note_commit.rs
@@ -1,18 +1,14 @@
 //! Note commitment logic for the Orchard circuit (Vanilla variation).
 
-use group::ff::PrimeField;
 use halo2_proofs::{
     circuit::{AssignedCell, Chip, Layouter},
-    plonk::{Advice, Column, ConstraintSystem, Error, Expression},
+    plonk::Error,
 };
 use pasta_curves::pallas;
 
 use crate::{
-    circuit::note_commit::{
-        DecomposeB, DecomposeD, DecomposeE, DecomposeG, DecomposeHVanilla, GdCanonicity,
-        PkdAssetCanonicity, PsiCanonicity, RhoCanonicity, ValueCanonicity, YCanonicity,
-    },
-    constants::{OrchardCommitDomains, OrchardFixedBases, OrchardHashDomains, T_P},
+    circuit::note_commit::{DecomposeB, DecomposeD, DecomposeE, DecomposeG, DecomposeHVanilla},
+    constants::{OrchardCommitDomains, OrchardFixedBases, OrchardHashDomains},
     value::NoteValue,
 };
 use halo2_gadgets::{
@@ -20,14 +16,8 @@ use halo2_gadgets::{
         chip::{EccChip, NonIdentityEccPoint},
         Point, ScalarFixed,
     },
-    sinsemilla::{
-        chip::{SinsemillaChip, SinsemillaConfig},
-        CommitDomain, Message, MessagePiece,
-    },
-    utilities::{
-        lookup_range_check::{PallasLookupRangeCheck, PallasLookupRangeCheckConfig},
-        RangeConstrained,
-    },
+    sinsemilla::{chip::SinsemillaChip, CommitDomain, Message, MessagePiece},
+    utilities::{lookup_range_check::PallasLookupRangeCheckConfig, RangeConstrained},
 };
 
 /*
@@ -43,158 +33,6 @@ use halo2_gadgets::{
         - psi is a base field element (255 bits).
 */
 
-#[allow(non_snake_case)]
-#[derive(Clone, Debug)]
-pub struct NoteCommitConfig<Lookup: PallasLookupRangeCheck> {
-    b: DecomposeB<Lookup>,
-    d: DecomposeD<Lookup>,
-    e: DecomposeE<Lookup>,
-    g: DecomposeG<Lookup>,
-    h: DecomposeHVanilla<Lookup>,
-    g_d: GdCanonicity<Lookup>,
-    pk_d_asset: PkdAssetCanonicity<Lookup>,
-    value: ValueCanonicity,
-    rho: RhoCanonicity<Lookup>,
-    psi: PsiCanonicity,
-    y_canon: YCanonicity,
-    advices: [Column<Advice>; 10],
-    sinsemilla_config:
-        SinsemillaConfig<OrchardHashDomains, OrchardCommitDomains, OrchardFixedBases, Lookup>,
-}
-
-#[derive(Clone, Debug)]
-pub struct NoteCommitChip<Lookup: PallasLookupRangeCheck> {
-    config: NoteCommitConfig<Lookup>,
-}
-
-impl<Lookup: PallasLookupRangeCheck> NoteCommitChip<Lookup> {
-    #[allow(non_snake_case)]
-    #[allow(clippy::many_single_char_names)]
-    pub(in crate::circuit) fn configure(
-        meta: &mut ConstraintSystem<pallas::Base>,
-        advices: [Column<Advice>; 10],
-        sinsemilla_config: SinsemillaConfig<
-            OrchardHashDomains,
-            OrchardCommitDomains,
-            OrchardFixedBases,
-            Lookup,
-        >,
-    ) -> NoteCommitConfig<Lookup> {
-        // Useful constants
-        let two = pallas::Base::from(2);
-        let two_pow_2 = pallas::Base::from(1 << 2);
-        let two_pow_4 = two_pow_2.square();
-        let two_pow_5 = two_pow_4 * two;
-        let two_pow_6 = two_pow_5 * two;
-        let two_pow_8 = two_pow_4.square();
-        let two_pow_9 = two_pow_8 * two;
-        let two_pow_10 = two_pow_9 * two;
-        let two_pow_58 = pallas::Base::from(1 << 58);
-        let two_pow_130 = Expression::Constant(pallas::Base::from_u128(1 << 65).square());
-        let two_pow_140 = Expression::Constant(pallas::Base::from_u128(1 << 70).square());
-        let two_pow_249 = pallas::Base::from_u128(1 << 124).square() * two;
-        let two_pow_250 = two_pow_249 * two;
-        let two_pow_254 = pallas::Base::from_u128(1 << 127).square();
-
-        let t_p = Expression::Constant(pallas::Base::from_u128(T_P));
-
-        // Columns used for MessagePiece and message input gates.
-        let col_l = advices[6];
-        let col_m = advices[7];
-        let col_r = advices[8];
-        let col_z = advices[9];
-
-        let b = DecomposeB::configure(meta, col_l, col_m, col_r, two_pow_4, two_pow_5, two_pow_6);
-        let d = DecomposeD::configure(meta, col_l, col_m, col_r, two, two_pow_2, two_pow_10);
-        let e = DecomposeE::configure(meta, col_l, col_m, col_r, two_pow_6);
-        let g = DecomposeG::configure(meta, col_l, col_m, two, two_pow_10);
-        let h = DecomposeHVanilla::configure(meta, col_l, col_m, col_r, two_pow_5);
-
-        let g_d = GdCanonicity::configure(
-            meta,
-            col_l,
-            col_m,
-            col_r,
-            col_z,
-            two_pow_130.clone(),
-            two_pow_250,
-            two_pow_254,
-            t_p.clone(),
-        );
-
-        let pk_d_asset = PkdAssetCanonicity::configure(
-            meta,
-            col_l,
-            col_m,
-            col_r,
-            col_z,
-            two_pow_4,
-            two_pow_140.clone(),
-            two_pow_254,
-            t_p.clone(),
-        );
-
-        let value =
-            ValueCanonicity::configure(meta, col_l, col_m, col_r, col_z, two_pow_8, two_pow_58);
-
-        let rho = RhoCanonicity::configure(
-            meta,
-            col_l,
-            col_m,
-            col_r,
-            col_z,
-            two_pow_4,
-            two_pow_140,
-            two_pow_254,
-            t_p.clone(),
-        );
-
-        let psi = PsiCanonicity::configure(
-            meta,
-            col_l,
-            col_m,
-            col_r,
-            col_z,
-            two_pow_9,
-            two_pow_130.clone(),
-            two_pow_249,
-            two_pow_254,
-            t_p.clone(),
-        );
-
-        let y_canon = YCanonicity::configure(
-            meta,
-            advices,
-            two,
-            two_pow_10,
-            two_pow_130,
-            two_pow_250,
-            two_pow_254,
-            t_p,
-        );
-
-        NoteCommitConfig {
-            b,
-            d,
-            e,
-            g,
-            h,
-            g_d,
-            pk_d_asset,
-            value,
-            rho,
-            psi,
-            y_canon,
-            advices,
-            sinsemilla_config,
-        }
-    }
-
-    pub(in crate::circuit) fn construct(config: NoteCommitConfig<Lookup>) -> Self {
-        Self { config }
-    }
-}
-
 pub(in crate::circuit) mod gadgets {
 
     use super::*;
@@ -202,6 +40,7 @@ pub(in crate::circuit) mod gadgets {
     use crate::circuit::note_commit::gadgets::{
         canon_bitshift_130, pkd_asset_x_canonicity, psi_canonicity, rho_canonicity, y_canonicity,
     };
+    use crate::circuit::note_commit::{NoteCommitChip, SpecificConfigForCircuit};
 
     #[allow(clippy::many_single_char_names)]
     #[allow(clippy::type_complexity)]
@@ -218,6 +57,13 @@ pub(in crate::circuit) mod gadgets {
         psi: AssignedCell<pallas::Base, pallas::Base>,
         rcm: ScalarFixed<pallas::Affine, EccChip<OrchardFixedBases>>,
     ) -> Result<Point<pallas::Affine, EccChip<OrchardFixedBases>>, Error> {
+        let vanilla_config = match &note_commit_chip.config.specific_config_for_circuit {
+            SpecificConfigForCircuit::Zsa(_) => {
+                panic!("Incorrect NoteCommitChip")
+            }
+            SpecificConfigForCircuit::Vanilla(specific_config) => specific_config.clone(),
+        };
+
         let lookup_config = chip.config().lookup_config();
 
         // `a` = bits 0..=249 of `x(g_d)`
@@ -367,7 +213,9 @@ pub(in crate::circuit) mod gadgets {
             .g
             .assign(&mut layouter, g, g_0, g_1.clone(), z1_g.clone())?;
 
-        let h_1 = cfg.h.assign(&mut layouter, h, h_0.clone(), h_1)?;
+        let h_1 = vanilla_config
+            .h_vanilla
+            .assign(&mut layouter, h, h_0.clone(), h_1)?;
 
         cfg.g_d
             .assign(&mut layouter, g_d, a, b_0, b_1, a_prime, z13_a, z13_a_prime)?;
@@ -416,10 +264,10 @@ pub(in crate::circuit) mod gadgets {
 mod tests {
     use core::iter;
 
-    use super::NoteCommitConfig;
     use crate::{
-        circuit::circuit_vanilla::note_commit::{gadgets, NoteCommitChip},
+        circuit::circuit_vanilla::note_commit::gadgets,
         circuit::gadget::assign_free_advice,
+        circuit::note_commit::{NoteCommitChip, NoteCommitConfig},
         constants::{
             fixed_bases::NOTE_COMMITMENT_PERSONALIZATION, OrchardCommitDomains, OrchardFixedBases,
             OrchardHashDomains, L_ORCHARD_BASE, L_VALUE, T_Q,
@@ -526,7 +374,7 @@ mod tests {
                     false,
                 );
                 let note_commit_config =
-                    NoteCommitChip::configure(meta, advices, sinsemilla_config);
+                    NoteCommitChip::configure(meta, advices, sinsemilla_config, false);
 
                 let ecc_config = EccChip::<OrchardFixedBases>::configure(
                     meta,
